Implementing DTO and ModelMapper

+++ src/main/java/com/example/springrestjpa/application/rest/EmployeeController.java
@@ -1,5 +1,7 @@
 package com.example.springrestjpa.application.rest;
 
+import com.example.springrestjpa.application.converter.EmployeeConverter;
+import com.example.springrestjpa.application.dto.EmployeeDTO;
 import com.example.springrestjpa.domain.Employee;
 import com.example.springrestjpa.domain.service.EmployeeService;
 import lombok.extern.slf4j.Slf4j;
@@ -9,7 +11,9 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 @RestController
@@ -23,46 +27,61 @@ public class EmployeeController {
     @Autowired
     private EmployeeService employeeService;
 
+    @Autowired
+    private EmployeeConverter employeeConverter;
+
     @GetMapping
-    public ResponseEntity<List<Employee>> getEmployees() {
-        return ResponseEntity.ok(employeeService.retrieveEmployees());
+    public ResponseEntity<List<EmployeeDTO>> getEmployees() {
+        List<Employee> results = employeeService.retrieveEmployees();
+
+        return ResponseEntity.ok(results.parallelStream()
+                .map(employeeConverter::convertToDTO)
+                .collect(Collectors.toList()));
     }
 
     @GetMapping(path = "/{employeeId}")
-    public ResponseEntity<Employee> getEmployee(@PathVariable(name = "employeeId") Long employeeId) {
+    public ResponseEntity<EmployeeDTO> getEmployee(@PathVariable(name = "employeeId") Long employeeId) {
         Employee employee = employeeService.getEmployee(employeeId);
         if (employee == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(employee);
+        return ResponseEntity.ok(employeeConverter.convertToDTO(employee));
     }
 
     @PostMapping
-    public ResponseEntity<Employee> saveEmployee(@RequestBody Employee employee) {
+    public ResponseEntity<EmployeeDTO> saveEmployee(@Valid @RequestBody EmployeeDTO employeeDTO) {
+        Employee employee = employeeConverter.convertToEntity(employeeDTO);
         Employee createdEmployee = employeeService.createEmployee(employee);
+
         log.info(NEW_EMPLOYEE_LOG, createdEmployee.toString());
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdEmployee);
+        return ResponseEntity.status(HttpStatus.CREATED).body(employeeConverter.convertToDTO(createdEmployee));
     }
 
     @PutMapping(path = "/{employeeId}")
-    public ResponseEntity<Employee> updateEmployee(@PathVariable(name = "employeeId") Long employeeId,
-                                   @RequestBody Employee employee) {
+    public ResponseEntity<EmployeeDTO> updateEmployee(@PathVariable(name = "employeeId") Long employeeId,
+                                   @Valid @RequestBody EmployeeDTO employeeDTO) {
         Employee empl = employeeService.getEmployee(employeeId);
         if (empl == null) {
-            ResponseEntity.notFound().build();
+            return ResponseEntity.notFound().build();
         }
 
+        Employee employee = employeeConverter.convertToEntity(employeeDTO);
         employee.setId(empl.getId());
         Employee updatedEmployee = employeeService.editEmployee(employee);
 
         log.info(EMPLOYEE_UPDATED_LOG, employeeId);
-
-        return ResponseEntity.ok(updatedEmployee);
+        return ResponseEntity.ok(employeeConverter.convertToDTO(updatedEmployee));
     }
 
     @DeleteMapping(path = "/{employeeId}")
     public ResponseEntity deleteEmployee(@PathVariable(name = "employeeId") Long employeeId) {
+        Employee empl = employeeService.getEmployee(employeeId);
+        if (empl == null) {
+            return ResponseEntity.notFound().build();
+        }
+
         employeeService.deleteEmployee(employeeId);
+
         log.info(EMPLOYEE_DELETED_LOG, employeeId);
         return ResponseEntity.ok().build();
     }
